+++ src/index.ts
@@ -33,3 +33,4 @@ export {
     buildSignedCredentialForExistingAccount,
 } from './credentialDeploymentTransactions';
 export { isAlias, getAlias } from './alias';
+export { deserializeContractState } from './deserialization';